+++ app/build.gradle
@@ -9,11 +9,11 @@ android {
 
     defaultConfig {
 
-        applicationId "dragosholban.com.androidpuzzlegame"
+        applicationId "de.stefan_oltmann.puzzleapp"
         minSdkVersion 16
         targetSdkVersion 30
         versionCode 1
-        versionName "1.0"
+        versionName "0.1"
     }
 
     buildTypes {
