+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="dragosholban.com.androidpuzzlegame">
+    package="de.stefan_oltmann.puzzleapp">
 
     <uses-permission
         android:name="android.permission.READ_EXTERNAL_STORAGE" />
@@ -14,6 +14,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         android:fullBackupContent="true">
+
         <activity
             android:name=".MainActivity"
             android:screenOrientation="portrait"
@@ -23,6 +24,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
         <activity android:name=".PuzzleActivity" />
 
         <provider
