+++ app/src/main/java/de/stefan_oltmann/puzzleapp/ImageAdapter.kt
@@ -1,4 +1,4 @@
-package dragosholban.com.androidpuzzlegame
+package de.stefan_oltmann.puzzleapp
 
 import android.content.Context
 import android.graphics.Bitmap
