+++ app/src/main/java/de/stefan_oltmann/puzzleapp/PuzzleActivity.kt
@@ -1,4 +1,4 @@
-package dragosholban.com.androidpuzzlegame
+package de.stefan_oltmann.puzzleapp
 
 import android.graphics.*
 import android.graphics.drawable.BitmapDrawable
