+++ app/src/main/res/layout/activity_puzzle.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="dragosholban.com.androidpuzzlegame.PuzzleActivity"
+    tools:context=".PuzzleActivity"
     android:background="@drawable/table_background">
 
     <ImageView
